Insight - WIP.

+++ src/main/java/core/service/oems/OEMSData.java
@@ -18,8 +18,9 @@ public class OEMSData extends SelfDescribingMarshallable {
     public double low;
     public double close;
     public double volume;
-    public double bid;
-    public double ask;
+    /*
+    public BigDecimal bid;
+    public BigDecimal ask;
     public BigDecimal bidSize;
     public BigDecimal askSize;
     public BigDecimal vwap;
@@ -27,6 +28,8 @@ public class OEMSData extends SelfDescribingMarshallable {
     public BigDecimal quoteVolume;
     public Date timestamp;
     public BigDecimal last;
+
+     */
     public CharSequence start;
     public CharSequence end;
